services/voice-agent/pyproject.toml
[project]
name = "syntera-voice-agent"
version = "0.1.0"
description = "Python-based LiveKit Agents service for Syntera voice AI"
requires-python = ">=3.9"
dependencies = [
    "livekit-agents[silero,turn-detector]>=1.2.0",
    "livekit-plugins-openai>=0.1.0",
    "livekit-plugins-noise-cancellation>=0.2.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "aiohttp>=3.9.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]






